requirements.txt
# make a venv and do: pip install -r requirements.txt
flask>=2.0.0
requests>=2.0.0
python-dotenv>=1.0.0


